+++ icons.js
@@ -1,0 +1,41 @@
+module.export = {
+    'rar': 'file-archive',
+    'zip': 'file-archive',
+    '7z': 'file-archive',
+    'tar': 'file-archive',
+
+    'doc': 'file-word',
+    'docx': 'file-word',
+    'xls': 'file-excel',
+    'xlsx': 'file-excel',
+    'ppt': 'file-powerpoint',
+    'pptx': 'file-powerpoint',
+    'pdf': 'file-pdf',
+    
+    'img': 'file-image',
+    'png': 'file-image',
+    'gif': 'file-image',
+    'jpg': 'file-image',
+    'jpeg': 'file-image',
+
+    'aac': 'file-audio',
+    'mp3': 'file-audio',
+    'wav': 'file-sound',
+
+    'mp4': 'file-video',
+    'mov': 'file-video',
+    'flv': 'file-video',
+
+    'txt': 'file-text',
+
+    'java': 'file-code',
+    'cs': 'file-code',
+    'js': 'file-code',
+    'jsp': 'file-code',
+    'css': 'file-code',
+    'sass': 'file-code',
+    'scss': 'file-code',
+    'ts': 'file-code',
+    'coffee': 'file-code',
+    'jsx': 'file-code'
+};