+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.0.4
+-------
+- Separated `icons.js` file
+- Refactored for ES6 function
+
 ### 0.0.3
 -------
 - Added new icons